added automatic folder creation + included it in chart and csv checking + readme update

+++ scripts/0_support/folders_management.R
@@ -1,6 +1,13 @@
 # CLEAR THE REPOSITORIES
 library(futile.logger)
 
+create_folder_if_not_exists <- function(path) {
+  if (!dir.exists(path)) {
+    dir.create(path, recursive = TRUE)
+    flog.info(paste("Create folder:", path))
+  }
+}
+
 # Data preparation
 clear_all <- function(country,
                       chart_path) {
@@ -12,19 +19,21 @@ clear_all <- function(country,
     flog.info(paste("Empty all the country folders"))
   } else {
     countries <- country
-    flog.info(paste("Emptying the folder: ", country))
   }
 
   for (country_chart in countries) {
     # Output charts
     outputpath <- paste(chart_path, "/", country_chart, "/", sep = "")
-    tryCatch(
-      {
-        unlink(paste0(outputpath, "*"))
-      },
-      error = function(e) {
-        flog.error(paste("Error clearing the folder: ", country_chart), e)
-      }
-    )
+    if (dir.exists(outputpath)){
+      tryCatch(
+        {
+          unlink(paste0(outputpath, "*"))
+          flog.info(paste("Empty the folder: ", outputpath))
+        },
+        error = function(e) {
+          flog.error(paste("Error clearing the folder: ", country_chart), e)
+        }
+      )
+    }
   }
 }
